Modified   Project/Networking/Binaries/Resources/InitialData_Default.json Modified   Project/Networking/Server/Data/GeneralData.cs Modified   Project/Networking/Server/OneByBotRoom.cs

+++ Project/Networking/Server/Data/GeneralData.cs
@@ -14,13 +14,13 @@ namespace Networking.Server.Data
 			return obj.Get<uint>("WaitForRestoreSession");
 		}
 
-		public static float GetChanceOfWhiteBot(int SplitTestGroupID)
+		public static uint GetChanceOfWhiteBot(int SplitTestGroupID)
 		{
 			ISerializeObject obj = GetGeneralObject(SplitTestGroupID);
 			if (obj == null)
 				return 0;
 
-			return obj.Get<float>("ChanceOfWhiteBot");
+			return obj.Get<uint>("ChanceOfWhiteBot");
 		}
 
 		public static uint GetStartGameDelay(int SplitTestGroupID)
